6.10.1-r1023_gdoadm.php

+++ Method/Cancel.php
@@ -7,15 +7,23 @@ use GDO\User\GDO_User;
 
 final class Cancel extends MethodPayment
 {
+    public function isTrivial() { return false; }
+    
 	public function execute()
 	{
 		$user = GDO_User::current();
-		$order = $this->getOrder();
+		if (!($order = $this->getOrder()))
+		{
+		    return $this->error('err_order');
+		}
+		
 		if ( (!$order->isPaid()) && ($order->getCreator()===$user) )
 		{
 			$order->delete();
+    		return $this->message('msg_order_cancelled')->add(Website::redirect($order->href_failure()));
 		}
-		return $this->message('msg_order_cancelled')->add(Website::redirect($order->href_failure()));
+
+		return $this->error('err_order_cancel');
 	}
 	
 }
